step 1: variable rearrangemnt
